% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_inference.R
\docType{data}
\name{genesets}
\alias{genesets}
\title{genesets}
\format{
An object of class \code{list} of length 2.
}
\usage{
genesets
}
\description{
A list consists of pathway genesets of Homo sapiens (human) and Mus musculus (house mouse).
}
\keyword{datasets}
