% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_inference.R
\docType{data}
\name{goolam}
\alias{goolam}
\title{goolam}
\format{
An object of class \code{SingleCellExperiment} with 41336 rows and 124 columns.
}
\usage{
goolam
}
\description{
goolam datas set save as a SCE object.
}
\keyword{datasets}
